serverless/requirements.txt
boto3
pillow
requests
